# Reduce calls to getDimensionsFromName by one for each datapoint