Reduce calls to getDimensionsFromName by one for each datapoint and makes things a bit more efficient

These numbers are after the PR  for making the  getDimensionsByName improvements and do not include them.
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkNewDatapoint-8     22056         17297         -21.58%

benchmark                   old allocs     new allocs     delta
BenchmarkNewDatapoint-8     60             47             -21.67%

benchmark                   old bytes     new bytes     delta
BenchmarkNewDatapoint-8     4336          3088          -28.78%
```