@mpetazzoni 2016-12-03
So are we not getting dimensions out of `type_instance` anymore?