@dgriff-tools 2016-12-03
It looks like this change moves the getting dimensions out of type_instance into the new function pointTypeInstance.