@mpetazzoni 2016-12-04
That's pretty confusing in terms of code organization. Why wouldn't the same logic apply to getting dimensions out of `plugin_instance`? If the point is to avoid allocations because dimensions are most likely not in `type_instance`, can't apply the same logic for `plugin_instance`, and ultimately, `metric`?