@mdubbyap 2016-12-05
No, type_instance becomes part of the metric name, plugin_instance and host do not.  I'll try to make this less confusing.