@mpetazzoni 2016-12-05
Looks a lot more readable, LGTM. I'll let @tedoc2000 merge when happy.